+++ lib/active_record/connection_adapters/postgis/spatial_table_definition.rb
@@ -50,7 +50,7 @@ module ActiveRecord  # :nodoc:
           spatial_type = geo_type(type)
           spatial_type << "Z" if options[:has_z]
           spatial_type << "M" if options[:has_m]
-          spatial_type << ",#{ options[:srid] || 4326 }"
+          spatial_type << ",#{options[:srid] || 4326}"
           spatial_type
         end
 
@@ -61,12 +61,12 @@ module ActiveRecord  # :nodoc:
             value << self.class.geo_type(spatial_type)
             value << "Z" if has_z?
             value << "M" if has_m?
-            value << ",#{ srid }"
+            value << ",#{srid}"
           end
         end
 
         def self.geo_type(type = "GEOMETRY")
-          g_type = type.to_s.gsub("_", "").upcase
+          g_type = type.to_s.delete("_").upcase
           return "POINT" if g_type == "STPOINT"
           return "POLYGON" if g_type == "STPOLYGON"
           g_type
@@ -116,7 +116,6 @@ module ActiveRecord  # :nodoc:
           @has_m = !!value
         end
       end
-
     end
   end
 end