+++ test/setup_test.rb
@@ -1,9 +1,16 @@
-require 'test_helper'
+require "test_helper"
 
 class SpatialQueriesTest < ActiveSupport::TestCase  # :nodoc:
-
   def test_ignore_tables
-    assert_equal %w(geometry_columns spatial_ref_sys layer topology), ::ActiveRecord::SchemaDumper.ignore_tables
+    expect_to_ignore = %w(
+      geography_columns
+      geometry_columns
+      layer
+      raster_columns
+      raster_overviews
+      spatial_ref_sys
+      topology
+    )
+    assert_equal expect_to_ignore, ::ActiveRecord::SchemaDumper.ignore_tables
   end
-
 end