+++ test/test_helper.rb
@@ -1,11 +1,11 @@
-require 'minitest/autorun'
-require 'minitest/pride'
-require 'mocha/mini_test'
-require 'active_record'
-require 'activerecord-postgis-adapter'
+require "minitest/autorun"
+require "minitest/pride"
+require "mocha/mini_test"
+require "active_record"
+require "activerecord-postgis-adapter"
 
 begin
-  require 'byebug'
+  require "byebug"
 rescue LoadError
   # ignore
 end
@@ -13,7 +13,7 @@ end
 class ActiveSupport::TestCase
   self.test_order = :sorted
 
-  DATABASE_CONFIG_PATH = File.dirname(__FILE__) << '/database.yml'
+  DATABASE_CONFIG_PATH = File.dirname(__FILE__) << "/database.yml"
 
   class SpatialModel < ActiveRecord::Base
     establish_connection YAML.load_file(DATABASE_CONFIG_PATH)