+++ History.md
@@ -1,3 +1,46 @@
+### 4.0.2 / 2016-11-13
+
+* Revert #237
+
+### 4.0.1 / 2016-11-08 *** YANKED
+
+* Auto-load tasks (#237)
+
+### 4.0.0 / 2016-06-30
+
+* Support ActiveRecord 5.0 (#213)
+* Fix schema dump null issues in JRuby (#229)
+
+### 3.1.4 / 2016-2-7
+
+* Ignore PostGIS views on schema dump - #208
+
+### 3.1.3 / 2016-1-15
+
+* Restrict ActiveRecord support to 4.2. See 649707cdf
+
+### 3.1.2 / 2015-12-29
+
+* Require rgeo-activerecord 4.0.4
+
+### 3.1.1 / 2015-12-28
+
+* Fix require for rgeo-activerecord 4.0.2
+* Rubocop-related cleanup #203
+
+### 3.1.0 / 2015-11-19
+
+* Add JRuby support (#199)
+
+### 3.0.0 / 2015-05-25
+
+* Support & require ActiveRecord 4.2 (#145)
+* Require rgeo-activerecord 4.0 (#180, 089d2dedd9b)
+* Rename adapter module from PostGISAdapter to PostGIS (c2fa909bb)
+* Breaking change: remove #set_rgeo_factory_settings
+* Breaking change: remove #rgeo_factory_for_column
+* Breaking change: remove #has_spatial_constraints?
+
 ### 2.2.1 / 2014-09-22
 
 * Update gemspec to not allow update to ActiveRecord 4.2, as it does not work.
@@ -5,12 +48,12 @@
 ### 2.2.0 / 2014-08-11
 
 * Add JRuby support
-  (https://github.com/rgeo/activerecord-postgis-adapter/pull/102)  
+  (https://github.com/rgeo/activerecord-postgis-adapter/pull/102)
 
 ### 2.1.1 / 2014-06-17
 
 * Correct behavior of non-geographic null: false columns
-  (https://github.com/rgeo/activerecord-postgis-adapter/pull/127)  
+  (https://github.com/rgeo/activerecord-postgis-adapter/pull/127)
 * Loosen rgeo-activerecord dependency
 
 ### 2.1.0 / 2014-06-11